+++ HomeReadingLibraryWeb/ClientApp/src/app/app.module.ts
@@ -40,6 +40,8 @@ import { MissingCheckinsReportComponent } from './components/missing-checkins-re
 import { YearEndCheckinsReportComponent } from './components/year-end-checkins-report/year-end-checkins-report.component';
 import { DatabaseRefreshComponent } from './components/database-refresh/database-refresh.component';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { StudentUploadComponent } from './components/student-upload/student-upload.component';
+import { VolunteerUploadComponent } from './components/volunteer-upload/volunteer-upload.component';
 
 @NgModule({
   declarations: [
@@ -65,7 +67,9 @@ import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
     AdminReportsComponent,
     MissingCheckinsReportComponent,
     YearEndCheckinsReportComponent,
-    DatabaseRefreshComponent
+    DatabaseRefreshComponent,
+    StudentUploadComponent,
+    VolunteerUploadComponent
   ],
   imports: [
     BrowserModule,