+++ HomeReadingLibraryWeb/ClientApp/src/app/modules/app-routing/app-routing.module.ts
@@ -17,6 +17,8 @@ import { AdminReportsComponent } from 'src/app/components/admin-reports/admin-re
 import { MissingCheckinsReportComponent } from 'src/app/components/missing-checkins-report/missing-checkins-report.component';
 import { YearEndCheckinsReportComponent } from 'src/app/components/year-end-checkins-report/year-end-checkins-report.component';
 import { DatabaseRefreshComponent } from 'src/app/components/database-refresh/database-refresh.component';
+import { StudentUploadComponent } from 'src/app/components/student-upload/student-upload.component';
+import { VolunteerUploadComponent } from 'src/app/components/volunteer-upload/volunteer-upload.component';
 
 
 const routes: Routes = [
@@ -35,6 +37,8 @@ const routes: Routes = [
             { path: 'adminreports', component: AdminReportsComponent, canActivate: [AuthGuard]},
             { path: 'missingcheckins', component: MissingCheckinsReportComponent, canActivate: [AuthGuard]},
             { path: 'yearendcheckins', component: YearEndCheckinsReportComponent, canActivate: [AuthGuard]},
+            { path: 'uploadstudents', component: StudentUploadComponent, canActivate: [AuthGuard]},
+            { path: 'uploadvolunteers', component: VolunteerUploadComponent, canActivate: [AuthGuard]},
             { path: 'databaserefresh', component: DatabaseRefreshComponent, canActivate: [AuthGuard]},
             { path: '**', redirectTo: 'home' }
         ];